[refactor] uuid -> userId header 변경 및 소스정리
