+++ src/main/java/com/hhplus/concert/api/token/presentation/TokenController.java
@@ -1,10 +1,9 @@
 package com.hhplus.concert.api.token.presentation;
 
 import com.hhplus.concert.api.token.application.TokenService;
-import com.hhplus.concert.api.token.presentation.dto.request.TokenPostReqDTO;
-import com.hhplus.concert.api.token.presentation.dto.response.TokenPostResDTO;
+import com.hhplus.concert.api.token.presentation.dto.request.TokenReqDTO;
+import com.hhplus.concert.api.token.presentation.dto.response.TokenResDTO;
 import com.hhplus.concert.api.token.presentation.mapper.TokenMapper;
-import com.hhplus.concert.api.token.presentation.dto.response.TokenGetResDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -30,15 +29,15 @@ public class TokenController {
     // 유저 대기열 토큰 발급 API
     @Operation(summary = "유저 대기열 토큰 발급 API")
     @PostMapping("")
-    public ResponseEntity<TokenPostResDTO> postToken(@Valid @RequestBody TokenPostReqDTO reqVo) {
+    public ResponseEntity<TokenResDTO> postToken(@Valid @RequestBody TokenReqDTO reqVo) {
         return ResponseEntity.ok(
-            TokenMapper.toDtoFromPost(tokenService.saveToken(TokenMapper.toEntity(reqVo))));
+            TokenMapper.toDto(tokenService.createWaitingToken(TokenMapper.toEntity(reqVo))));
     }
 
     // 유저 대기열 토큰 조회 API
     @Operation(summary = "유저 대기열 토큰 조회 API")
-    @GetMapping("/{uuid}")
-    public ResponseEntity<TokenGetResDTO> getToken(@PathVariable String uuid) {
-        return ResponseEntity.ok(TokenMapper.toDtoFromGet(tokenService.getToken(uuid)));
+    @GetMapping("/{userId}")
+    public ResponseEntity<TokenResDTO> getToken(@PathVariable Long userId) {
+        return ResponseEntity.ok(TokenMapper.toDto(tokenService.getWaitingToken(userId)));
     }
 }
